refactor(web): separate someday and regular event sagas
this is to make it easier to understand sagas work
