+++ packages/web/src/ducks/events/sagas/event.sagas.ts
@@ -1,6 +1,6 @@
 import { normalize } from "normalizr";
 import dayjs from "dayjs";
-import { call, put, takeLatest, select } from "@redux-saga/core/effects";
+import { call, put, select } from "@redux-saga/core/effects";
 import { Params_Events, Schema_Event } from "@core/types/event.types";
 import { YEAR_MONTH_DAY_FORMAT } from "@core/constants/date.constants";
 import { RootState } from "@web/store";
@@ -15,7 +15,6 @@ import {
 } from "@web/common/utils/event.util";
 import { Schema_GridEvent } from "@web/common/types/web.event.types";
 import { ID_OPTIMISTIC_PREFIX } from "@web/common/constants/web.constants";
-import { AxiosResponse } from "axios";
 
 import {
   createEventSlice,
@@ -26,7 +25,6 @@ import {
 } from "../slices/event.slice";
 import { getWeekEventsSlice } from "../slices/week.slice";
 import {
-  Action_ConvertSomedayEvent,
   Action_ConvertTimedEvent,
   Action_CreateEvent,
   Action_DeleteEvent,
@@ -39,59 +37,13 @@ import {
   Entities_Event,
 } from "../event.types";
 import { getSomedayEventsSlice } from "../slices/someday.slice";
-import { Action_Someday_Reorder } from "../slices/someday.slice.types";
 import {
   insertOptimisticEvent,
   normalizedEventsSchema,
   replaceOptimisticId,
-} from "./event.saga.util";
+} from "./saga.util";
 
-function* convertSomedayEvent({ payload }: Action_ConvertSomedayEvent) {
-  const { _id, updatedFields } = payload;
-  const optimisticId: string | null = null;
-
-  try {
-    //get grid event from store
-    const currEvent = (yield select((state: RootState) =>
-      selectEventById(state, _id)
-    )) as Response_GetEventsSaga;
-    const gridEvent = { ...currEvent, ...updatedFields };
-    // remove extra props before sending to DB
-    delete gridEvent.order;
-    delete gridEvent.recurrence;
-
-    //get optimisitcGridEvent
-    const optimisticGridEvent = replaceIdWithOptimisticId(gridEvent);
-    yield put(getSomedayEventsSlice.actions.remove({ _id }));
-    yield* insertOptimisticEvent(optimisticGridEvent, false);
-
-    // call API
-    const response = (yield call(
-      EventApi.edit,
-      _id,
-      gridEvent,
-      {}
-    )) as AxiosResponse<Schema_Event>;
-
-    const convertedEvent = response.data;
-
-    // replace ids
-    yield* replaceOptimisticId(
-      optimisticGridEvent._id,
-      convertedEvent._id as string,
-      false
-    );
-  } catch (error) {
-    if (optimisticId) {
-      yield put(eventsEntitiesSlice.actions.delete({ _id: optimisticId }));
-    }
-    yield put(getSomedayEventsSlice.actions.insert(_id));
-    yield put(getSomedayEventsSlice.actions.error());
-    handleError(error as Error);
-  }
-}
-
-function* convertTimedEvent({ payload }: Action_ConvertTimedEvent) {
+export function* convertTimedEvent({ payload }: Action_ConvertTimedEvent) {
   try {
     const res = yield call(EventApi.edit, payload.event._id, payload.event);
     const event = res.data as Schema_Event;
@@ -122,7 +74,7 @@ function* convertTimedEvent({ payload }: Action_ConvertTimedEvent) {
   }
 }
 
-function* createEvent({ payload }: Action_CreateEvent) {
+export function* createEvent({ payload }: Action_CreateEvent) {
   const event = replaceIdWithOptimisticId(payload);
   const optimisticId = event._id;
 
@@ -167,18 +119,6 @@ export function* deleteEvent({ payload }: Action_DeleteEvent) {
   }
 }
 
-export function* deleteSomedayEvent({ payload }: Action_DeleteEvent) {
-  try {
-    yield put(eventsEntitiesSlice.actions.delete(payload));
-
-    yield call(EventApi.delete, payload._id);
-  } catch (error) {
-    yield put(getSomedayEventsSlice.actions.error());
-    handleError(error as Error);
-    yield put(getSomedayEventsSlice.actions.request());
-  }
-}
-
 export function* editEvent({ payload }: Action_EditEvent) {
   const { _id, applyTo, event, shouldRemove } = payload;
 
@@ -197,7 +137,7 @@ export function* editEvent({ payload }: Action_EditEvent) {
   }
 }
 
-function* getCurrentMonthEvents({ payload }: Action_GetPaginatedEvents) {
+export function* getCurrentMonthEvents({ payload }: Action_GetPaginatedEvents) {
   try {
     const startDate = dayjs().startOf("month").format(YEAR_MONTH_DAY_FORMAT);
     const endDate = dayjs().endOf("month").format(YEAR_MONTH_DAY_FORMAT);
@@ -242,31 +182,7 @@ function* getEvents(
   }
 }
 
-export function* getSomedayEvents({ payload }: Action_GetEvents) {
-  try {
-    const res = (yield call(EventApi.get, {
-      someday: true,
-      startDate: payload.startDate,
-      endDate: payload.endDate,
-    })) as Response_GetEventsSuccess;
-
-    const normalizedEvents = normalize<Schema_Event>(res.data, [
-      normalizedEventsSchema(),
-    ]);
-    yield put(
-      eventsEntitiesSlice.actions.insert(normalizedEvents.entities.events)
-    );
-
-    const data = {
-      data: normalizedEvents.result as Payload_NormalizedAsyncAction,
-    };
-    yield put(getSomedayEventsSlice.actions.success(data));
-  } catch (error) {
-    yield put(getSomedayEventsSlice.actions.error());
-  }
-}
-
-function* getWeekEvents({ payload }: Action_GetEvents) {
+export function* getWeekEvents({ payload }: Action_GetEvents) {
   try {
     const data = (yield call(getEvents, payload)) as Response_GetEventsSaga;
     yield put(getWeekEventsSlice.actions.success(data));
@@ -274,32 +190,4 @@ function* getWeekEvents({ payload }: Action_GetEvents) {
     yield put(getWeekEventsSlice.actions.error());
     handleError(error as Error);
   }
-}
-
-function* reorderSomedayEvents({ payload }: Action_Someday_Reorder) {
-  try {
-    yield call(EventApi.reorder, payload);
-  } catch (error) {
-    yield put(getSomedayEventsSlice.actions.error());
-    handleError(error as Error);
-  }
-}
-
-/************
- * Assemble
- ***********/
-export function* eventsSagas() {
-  yield takeLatest(getWeekEventsSlice.actions.request, getWeekEvents);
-  yield takeLatest(getWeekEventsSlice.actions.convert, convertTimedEvent);
-  yield takeLatest(
-    getCurrentMonthEventsSlice.actions.request,
-    getCurrentMonthEvents
-  );
-  yield takeLatest(getSomedayEventsSlice.actions.convert, convertSomedayEvent);
-  yield takeLatest(getSomedayEventsSlice.actions.request, getSomedayEvents);
-  yield takeLatest(getSomedayEventsSlice.actions.delete, deleteSomedayEvent);
-  yield takeLatest(getSomedayEventsSlice.actions.reorder, reorderSomedayEvents);
-  yield takeLatest(createEventSlice.actions.request, createEvent);
-  yield takeLatest(editEventSlice.actions.request, editEvent);
-  yield takeLatest(deleteEventSlice.actions.request, deleteEvent);
 }
